requirements.txt
Django==5.1.3
djangorestframework==3.15.2
dj-rest-auth==7.0.0
django-allauth==0.61.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
python-dotenv==1.0.1
gunicorn==23.0.0
requests==2.32.3
cryptography==44.0.0
